Refactor data types across multiple modules for improved consistency and performance
- Changed floating-point types from `f64` to `f32` in the `cell`, `sensor`, `evolve`, `gene`, `mutation`, `layer`, `network`, `neuron`, `object`, `simulator`, `statistics`, and various system modules.
- Updated function signatures and internal calculations to reflect the new type, enhancing performance and reducing memory usage.
- Ensured compatibility across the codebase by adjusting related methods and data structures.

These changes promote a more uniform approach to numerical data handling, facilitating future development and optimization.

+++ osmos-core/src/cell.rs
@@ -1,15 +1,15 @@
-const BASE_SPEED: f64 = 2.0;
+const BASE_SPEED: f32 = 2.0;
 
 pub struct Cell {
-    pub position: nalgebra::Point2<f64>,
-    pub direction: nalgebra::Vector2<f64>,
-    pub velocity: nalgebra::Vector2<f64>,
+    pub position: nalgebra::Point2<f32>,
+    pub direction: nalgebra::Vector2<f32>,
+    pub velocity: nalgebra::Vector2<f32>,
     pub energy: usize,
     pub sensor: crate::sensor::Sensor,
 }
 
 impl Cell {
-    pub fn random(rng: &mut impl rand::RngCore, max_x: f64, max_y: f64) -> Self {
+    pub fn random(rng: &mut impl rand::RngCore, max_x: f32, max_y: f32) -> Self {
         let sensor_range = nalgebra::Vector2::new(max_x, max_y).magnitude() * 0.5;
         let mut cell = Self {
             position: nalgebra::Point2::new(0.0, 0.0),
@@ -22,14 +22,14 @@ impl Cell {
         cell
     }
 
-    pub fn random_position(&mut self, rng: &mut impl rand::RngCore, max_x: f64, max_y: f64) {
+    pub fn random_position(&mut self, rng: &mut impl rand::RngCore, max_x: f32, max_y: f32) {
         let random_x = rand::Rng::gen_range(rng, 0.0..=max_x);
         let random_y = rand::Rng::gen_range(rng, 0.0..=max_y);
         self.position.x = random_x;
         self.position.y = random_y;
     }
 
-    pub fn get_speed(&self) -> f64 {
-        BASE_SPEED + (BASE_SPEED / self.energy as f64)
+    pub fn get_speed(&self) -> f32 {
+        BASE_SPEED + (BASE_SPEED / self.energy as f32)
     }
 }
