+++ osmos-ga/src/gene.rs
@@ -1,7 +1,7 @@
-pub type Gene = Vec<f64>;
+pub type Gene = Vec<f32>;
 
 pub trait GeneObject {
-    fn build(rng: &mut impl rand::RngCore, gene: Gene, id: usize, max_x: f64, max_y: f64) -> Self;
+    fn build(rng: &mut impl rand::RngCore, gene: Gene, id: usize, max_x: f32, max_y: f32) -> Self;
 
     fn gene(&self) -> Gene;
